+++ packages/polaris-viz/CHANGELOG.md
@@ -11,6 +11,10 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - `HorizontalBarChart` and `VerticalBarChart` don't display a bar if a passed value is null or missing.
 
+### Added
+
+- Render xAxis and dual yAxis annotations in `<ComboChart />`.
+
 ## [6.1.0] - 2022-07-21
 
 ### Removed